ALFMOB-58: Code review typo and codegen implementation details
